+++ packages/apalis-core/src/backend.rs
@@ -2,7 +2,6 @@ use std::{any::type_name, future::Future};
 
 use futures::Stream;
 use serde::{Deserialize, Serialize};
-use tower::Service;
 
 use crate::{
     poller::Poller,
@@ -15,18 +14,18 @@ use crate::{
 ///
 /// [`Storage`]: crate::storage::Storage
 /// [`MessageQueue`]: crate::mq::MessageQueue
-pub trait Backend<Req, Res> {
+pub trait Backend<Req> {
     /// The stream to be produced by the backend
     type Stream: Stream<Item = Result<Option<Req>, crate::error::Error>>;
 
     /// Returns the final decoration of layers
     type Layer;
 
+    /// The way data is stored in the backend
+    type Compact;
+
     /// Returns a poller that is ready for streaming
-    fn poll<Svc: Service<Req, Response = Res>>(
-        self,
-        worker: &Worker<Context>,
-    ) -> Poller<Self::Stream, Self::Layer>;
+    fn poll(self, worker: &Worker<Context>) -> Poller<Self::Stream, Self::Layer>;
 }
 
 /// Represents functionality that allows reading of jobs and stats from a backend