+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project are documented in this file.
 ## [Unreleased]
 
 ### Added
+- **api** add associated types to the Backend trait ([#516](https://github.com/geofmureithi/apalis/pull/516))
 - **retry layer**: Integrate retry logic with task handling ([#512](https://github.com/geofmureithi/apalis/pull/512))
 - **generic retry**: Persist check for tasks ([#498](https://github.com/geofmureithi/apalis/pull/498))
 - **native TLS**: Add `async-std-comp-native-tls` and `tokio-comp-native-tls` features ([#525](https://github.com/geofmureithi/apalis/pull/525))